requirements.txt
biopython==1.70
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==1.10.5
django-blastplus==0.4.0
django-developer-panel==0.1.2
django-haystack==2.6.1
django-rest-framework==0.1.0
djangorestframework==3.7.7
elasticsearch==2.4.1
numpy==1.14.0
urllib3==1.22
